backend/requirements.txt
--index-url=https://pypi.org/simple/
asgiref==3.7.2
gunicorn
anyio==4.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
packaging==23.2
psycopg2-binary==2.9.9
python-telegram-bot==20.6
pytz==2023.3.post1
requests==2.31.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
